# server/docs/swagger.yaml
basePath: /
definitions:
  main.FeedItem:
    properties:
      card_type:
        description: weather | safe_alert | cultural_tip | map_coord | article
        type: string
      data:
        additionalProperties: true
        type: object
      id:
        type: string
      priority:
        description: high | medium | low
        type: string
      timestamp:
        description: ISO 8601 timestamp
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the gateway service for the AiGuardian application.
  title: AiGuardian Gateway API
  version: "1.0"
paths:
  /api/agent/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an agent file (.m or .zip)
      parameters:
      - description: Agent File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload Agent
      tags:
      - agent
  /api/chat/stream:
    post:
      consumes:
      - application/json
      description: Send a message to an agent and stream the response
      parameters:
      - description: Chat Request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE Stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Chat with Agent
      tags:
      - chat
  /api/feed:
    delete:
      description: Clear the insight stream feed
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clear Feed
      tags:
      - feed
    get:
      description: Get the current insight stream feed
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.FeedItem'
            type: array
      summary: Get Feed
      tags:
      - feed
  /health:
    get:
      description: Get service health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health Check
      tags:
      - system
swagger: "2.0"